+++ pytorch_lightning/trainer/properties.py
@@ -491,8 +491,8 @@ class TrainerProperties(ABC):
     @property
     def evaluation_loop(self) -> EvaluationLoop:
         if self.state.fn in (TrainerFn.FITTING, TrainerFn.TUNING):
-            return self.fit_loop.val_loop
-        if self.state.fn == TrainerFn.VALIDATING:
+            return self.fit_loop.epoch_loop.val_loop
+        elif self.state.fn == TrainerFn.VALIDATING:
             return self.validation_loop
         if self.state.fn == TrainerFn.TESTING:
             return self.test_loop
