+++ tests/trainer/test_trainer.py
@@ -1101,9 +1101,9 @@ def test_num_sanity_val_steps(tmpdir, limit_val_batches):
     assert trainer.num_sanity_val_steps == num_sanity_val_steps
 
     with patch.object(
-        trainer.fit_loop.val_loop.epoch_loop,
+        trainer.fit_loop.epoch_loop.val_loop.epoch_loop,
         "evaluation_step",
-        wraps=trainer.fit_loop.val_loop.epoch_loop.evaluation_step
+        wraps=trainer.fit_loop.epoch_loop.val_loop.epoch_loop.evaluation_step
     ) as mocked:
         val_dataloaders = model.val_dataloader__multiple_mixed_length()
         trainer.fit(model, val_dataloaders=val_dataloaders)
@@ -1131,9 +1131,9 @@ def test_num_sanity_val_steps_neg_one(tmpdir, limit_val_batches):
     assert trainer.num_sanity_val_steps == float("inf")
 
     with patch.object(
-        trainer.fit_loop.val_loop.epoch_loop,
+        trainer.fit_loop.epoch_loop.val_loop.epoch_loop,
         "evaluation_step",
-        wraps=trainer.fit_loop.val_loop.epoch_loop.evaluation_step
+        wraps=trainer.fit_loop.epoch_loop.val_loop.epoch_loop.evaluation_step
     ) as mocked:
         val_dataloaders = model.val_dataloader__multiple()
         trainer.fit(model, val_dataloaders=val_dataloaders)
